+++ src/main/java/edu/wpi/cs3733/D22/teamC/controller/service_request/MedicalEquipmentSRCreateController.java
@@ -3,15 +3,12 @@ package edu.wpi.cs3733.D22.teamC.controller.service_request;
 import com.jfoenix.controls.*;
 import com.jfoenix.controls.datamodels.treetable.RecursiveTreeObject;
 import edu.wpi.cs3733.D22.teamC.entity.service_request.medical_equipment.MedicalEquipmentServiceRequest;
-import edu.wpi.cs3733.D22.teamC.entity.service_request.medical_equipment.MedicalEquipmentTable;
-import javafx.beans.value.ObservableValue;
+import edu.wpi.cs3733.D22.teamC.entity.service_request.medical_equipment.MedicalEquipmentSRTable;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
-import javafx.fxml.Initializable;
 import javafx.scene.control.*;
-import javafx.util.Callback;
 
 import java.net.URL;
 import java.util.ResourceBundle;
@@ -24,11 +21,11 @@ public class MedicalEquipmentSRCreateController extends ServiceRequestCreateCont
     @FXML private JFXComboBox<String> equipType;
 
     // For table
-    @FXML private JFXTreeTableView<MedicalEquipmentTable> table;
-    ObservableList<MedicalEquipmentTable> METList = FXCollections.observableArrayList();
-    final TreeItem<MedicalEquipmentTable> root = new RecursiveTreeItem<MedicalEquipmentTable>(METList, RecursiveTreeObject::getChildren);
+    @FXML private JFXTreeTableView<MedicalEquipmentSRTable> table;
+    ObservableList<MedicalEquipmentSRTable> METList = FXCollections.observableArrayList();
+    final TreeItem<MedicalEquipmentSRTable> root = new RecursiveTreeItem<MedicalEquipmentSRTable>(METList, RecursiveTreeObject::getChildren);
 
-    ObservableList<MedicalEquipmentTable> data;
+    ObservableList<MedicalEquipmentSRTable> data;
 
 
     @Override
@@ -41,71 +38,18 @@ public class MedicalEquipmentSRCreateController extends ServiceRequestCreateCont
         equipType.getItems().add("Portable X-Ray");
         equipType.getItems().add("Infusion Pumps (30)");
 
-        //Columns for table
-        JFXTreeTableColumn<MedicalEquipmentTable, String> IDCol = new JFXTreeTableColumn<>("Priority");
-        IDCol.setPrefWidth(80);
-        IDCol.setCellValueFactory(new Callback<TreeTableColumn.CellDataFeatures<MedicalEquipmentTable, String>, ObservableValue<String>>() {
-            @Override
-            public ObservableValue<String> call(TreeTableColumn.CellDataFeatures<MedicalEquipmentTable, String> param) {
-                return param.getValue().getValue().priorityProperty();
-            }
-        });
-        JFXTreeTableColumn<MedicalEquipmentTable, String> assigneeCol = new JFXTreeTableColumn<>("Assignee");
-        assigneeCol.setPrefWidth(80);
-        assigneeCol.setCellValueFactory(new Callback<TreeTableColumn.CellDataFeatures<MedicalEquipmentTable, String>, ObservableValue<String>>() {
-            @Override
-            public ObservableValue<String> call(TreeTableColumn.CellDataFeatures<MedicalEquipmentTable, String> param) {
-                return param.getValue().getValue().assigneeIDProperty();
-            }
-        });
-        JFXTreeTableColumn<MedicalEquipmentTable, String> statusCol = new JFXTreeTableColumn<>("Status");
-        statusCol.setPrefWidth(80);
-        statusCol.setCellValueFactory(new Callback<TreeTableColumn.CellDataFeatures<MedicalEquipmentTable, String>, ObservableValue<String>>() {
-            @Override
-            public ObservableValue<String> call(TreeTableColumn.CellDataFeatures<MedicalEquipmentTable, String> param) {
-                return param.getValue().getValue().statusProperty();
-            }
-        });
-        JFXTreeTableColumn<MedicalEquipmentTable, String> locationCol = new JFXTreeTableColumn<>("Location");
-        locationCol.setPrefWidth(80);
-        locationCol.setCellValueFactory(new Callback<TreeTableColumn.CellDataFeatures<MedicalEquipmentTable, String>, ObservableValue<String>>() {
-            @Override
-            public ObservableValue<String> call(TreeTableColumn.CellDataFeatures<MedicalEquipmentTable, String> param) {
-                return param.getValue().getValue().locationProperty();
-            }
-        });
-        JFXTreeTableColumn<MedicalEquipmentTable, String> typeCol = new JFXTreeTableColumn<>("Type");
-        typeCol.setPrefWidth(80);
-        typeCol.setCellValueFactory(new Callback<TreeTableColumn.CellDataFeatures<MedicalEquipmentTable, String>, ObservableValue<String>>() {
-            @Override
-            public ObservableValue<String> call(TreeTableColumn.CellDataFeatures<MedicalEquipmentTable, String> param) {
-                return param.getValue().getValue().equipmentTypeProperty();
-            }
-        });
-        JFXTreeTableColumn<MedicalEquipmentTable, String> typeIDCol = new JFXTreeTableColumn<>("EquipID");
-        typeIDCol.setPrefWidth(80);
-        typeIDCol.setCellValueFactory(new Callback<TreeTableColumn.CellDataFeatures<MedicalEquipmentTable, String>, ObservableValue<String>>() {
-            @Override
-            public ObservableValue<String> call(TreeTableColumn.CellDataFeatures<MedicalEquipmentTable, String> param) {
-                return param.getValue().getValue().equipmentIDProperty();
-            }
-        });
+        MedicalEquipmentSRTable.createTableColumns(table);
+        table.setRoot(root);
+        table.setShowRoot(false);
 
         //Practice classes to add
-        MedicalEquipmentTable met1 = new MedicalEquipmentTable("Bed", "15", "123456",
+        MedicalEquipmentSRTable met1 = new MedicalEquipmentSRTable("Bed", "15", "123456",
               "Room 202", "Done", "High");
-        MedicalEquipmentTable met2 = new MedicalEquipmentTable("Infusion Pump", "35", "392843",
+        MedicalEquipmentSRTable met2 = new MedicalEquipmentSRTable("Infusion Pump", "35", "392843",
                "Room 305", "Blank", "Low");
         METList.add(met1);
         METList.add(met2);
-
-        //Sets columns
-        table.getColumns().setAll(IDCol, assigneeCol, statusCol, locationCol, typeCol, typeIDCol);
-
-        table.setRoot(root);
-        table.setShowRoot(false);
     }
-
 
     @FXML
     void clickReset(ActionEvent event) {
@@ -133,13 +77,14 @@ public class MedicalEquipmentSRCreateController extends ServiceRequestCreateCont
         int type = medEquip.getEquipEnum(equipType.getValue());
         String num = equipID.getText();
         medEquip.setEquipmentID(type + num);
-        medEquip.setMet();
         clickReset(event);
 
-        METList.add(medEquip.getMet());
-        System.out.println(medEquip.getAssigneeID());
-        System.out.println(medEquip.getStatus());
-        System.out.println(medEquip.getEquipmentID());
+
+        // Table Entry
+        MedicalEquipmentSRTable met = new MedicalEquipmentSRTable(medEquip);
+        METList.add(met);
+
+
         return medEquip;
     }
 }
